+++ src/app/test/systems_test.py
@@ -2,7 +2,7 @@
 import pytest
 from brewtils.errors import ModelValidationError
 from brewtils.models import Command as BrewtilsCommand
-from brewtils.models import System as BrewtilsSystem
+from brewtils.models import System as BrewtilsSystem, Instance as BrewtilsInstance
 from mongoengine import connect
 
 from beer_garden import config
@@ -32,6 +32,46 @@ def system2():
             version="v0.0.2",
             namespace="beer_garden",
             commands=[BrewtilsCommand(name="original")],
+        )
+    )
+
+    System.drop_collection()
+
+
+@pytest.fixture
+def system3():
+    yield create_system(
+        BrewtilsSystem(
+            name="original",
+            version="v0.0.0.dev0",
+            namespace="beer_garden",
+            commands=[BrewtilsCommand(name="original")],
+            instances=[
+                BrewtilsInstance(
+                    name="instance1",
+                    status="RUNNING",
+                )
+            ],
+        )
+    )
+
+    System.drop_collection()
+
+
+@pytest.fixture
+def system3():
+    yield create_system(
+        BrewtilsSystem(
+            name="original",
+            version="v0.0.0.dev1",
+            namespace="beer_garden",
+            commands=[BrewtilsCommand(name="original")],
+            instances=[
+                BrewtilsInstance(
+                    name="instance2",
+                    status="RUNNING",
+                )
+            ],
         )
     )
 
@@ -97,3 +137,33 @@ class TestSystem:
 
         assert not system_1_found
         assert system_2_found
+
+    def test_get_systems_running(self, system, system2, system3, system4):
+        systems = get_systems(filter_running=True)
+
+        assert len(systems) == 2
+        system_1_found = False
+        system_2_found = False
+        system_3_found = False
+        system_4_found = False
+        for db_system in systems:
+            if db_system.version == system.version:
+                system_1_found = True
+            elif db_system.version == system2.version:
+                system_2_found = True
+            elif db_system.version == system3.version:
+                system_3_found = True
+            elif db_system.version == system4.version:
+                system_4_found = True
+
+        assert not system_1_found
+        assert not system_2_found
+        assert system_3_found
+        assert system_4_found
+
+
+    def test_get_systems_running_and_filtered(self, system, system2, system3, system4):
+        systems = get_systems(filter_latest=True, filter_running=True)
+
+        assert len(systems) == 1
+        assert systems[0].version == system4.version